agent page

+++ modules/koki-portal/src/main/kotlin/com/wutsi/koki/portal/agent/model/AgentMetricModel.kt
@@ -3,7 +3,7 @@ package com.wutsi.koki.portal.agent.model
 import com.wutsi.koki.portal.common.model.MoneyModel
 
 data class AgentMetricModel(
-    val total: Int = 0,
+    val total: Long = 0,
     val minPrice: MoneyModel = MoneyModel(),
     val maxPrice: MoneyModel = MoneyModel(),
     val averagePrice: MoneyModel = MoneyModel(),
